It was moved to the VC.
Adding it here, would forced to pass the VC in the init of the VM. Because we are using storyboards. Either that, or get the rootVC via the UIApplication.sharedApplication. At the end of the day, in order to push something (pushViewController), the action needs to start from a VC as well. Having the VC as the tableViewDelegate, circumvents all these problems and then some.