Abstracted ViewModel of CocoaAction
Since CocoaAction is directly related with the user interface, in theory
seems more logical keeping the ViewModel only responsible for provide an
action that can be used by the interface which in turn can opt in for
doing the wrapping.

By exposing the Action itself we also have more flexibility to handle
errors resulting of the execution of the action.